requirements.txt
# Flask Framework
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1

# Database (SQLite - no extra dependencies needed)
SQLAlchemy==2.0.23

# Data Processing (for CSV parsing)
python-dateutil==2.8.2

# Environment Variables
python-dotenv==1.0.0

# HTTP Requests (for Groq AI API calls)
requests==2.31.0

# Production Server
gunicorn==21.2.0

# Utilities
werkzeug==3.0.1

# Authentication
Flask-JWT-Extended==4.6.0
bcrypt==4.1.2

# PDF Processing
pdfplumber==0.10.3
pikepdf==10.1.0

# Excel Processing
openpyxl==3.1.2
xlrd==2.0.1
